% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_add_columns.R,
%   R/wrangle_data_manipulation.R
\name{add_fgr_nrwbund_ltxt}
\alias{add_fgr_nrwbund_ltxt}
\title{Add column \code{fgr_nrwbund_ltxt} to data frame}
\usage{
add_fgr_nrwbund_ltxt(df, df_orga)

add_fgr_nrwbund_ltxt(df, df_orga)
}
\arguments{
\item{df}{Data frame containing \code{fgr_nrwbund_id}}

\item{df_orga}{Data frame containing \code{fgr_nrwbund_id} and \code{fgr_nrwbund_ltxt}}

\item{df}{Data frame containing \code{fgr_nrwbund_id}}

\item{df_orga}{Data frame containing \code{fgr_nrwbund_id} and \code{fgr_nrwbund_ltxt}}
}
\value{
Data frame with additional column \code{fgr_nrwbund_ltxt}

Data frame with additional column \code{fgr_nrwbund_ltxt}
}
\description{
Add column \code{fgr_nrwbund_ltxt} to data frame

Add column \code{fgr_nrwbund_ltxt} to data frame
}
\examples{
add_fgr_nrwbund_ltxt(df)
add_fgr_nrwbund_ltxt(df)
}
