% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_add_columns.R
\name{add_figure_count}
\alias{add_figure_count}
\title{Adds column \code{figure_count} to data frame}
\usage{
add_figure_count(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with additional column \code{figure_count}
}
\description{
Adds column \code{figure_count} to data frame
}
\examples{
\dontrun{
add_figure_count(df)
}
}
