% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_add_columns.R
\name{add_file_name}
\alias{add_file_name}
\title{Add column \code{file_name} to data frame}
\usage{
add_file_name(df)
}
\arguments{
\item{df}{}
}
\value{
Data frame with additional column \code{file_name}
}
\description{
Add column \code{file_name} to data frame
}
\examples{
add_file_name(df)
}
