% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_add_columns.R
\name{add_label_position}
\alias{add_label_position}
\title{Add centered y-coordinates for filtered data labels of figure types 1 and 2}
\usage{
add_label_position(
  df,
  x_var,
  y_var,
  facet_var = NULL,
  fill_var,
  filter_cutoff = 0.04,
  fill_reverse = FALSE,
  is_percentage = FALSE
)
}
\arguments{
\item{df}{Data frame}

\item{x_var}{Required variable name for the variable containing the discrete
x-coordinates.}

\item{y_var}{Required variable name for the variable containing the
continuous y-coordinates.}

\item{facet_var}{Optional variable name for the discrete variable to facet
by, defaults to NULL.}

\item{fill_var}{Variable name for the discrete variable which determines the
groups to be stacked, e.g. degree.}

\item{filter_cutoff}{Optional cutoff value for the suppression of data
labels. By default, all values below 0.04 of the total value of the
stacked bar chart are suppressed.}

\item{fill_reverse}{Boolean indicating whether the order of the fill variable
should be reversed, defaults to FALSE.}

\item{is_percentage}{Optional boolean indicating whether the value label is
expressed in absolute numbers or as a percentage, defaults to false}
}
\value{
Data frame with additional column "label_" + the name of the
   y-coordinate variable
}
\description{
Calculates the y-coordinates for stacked bar charts, centered for
   each group. The default function does not work if some labels are
   filtered.
}
\examples{
\dontrun{
add_label_position(df, x_var, y_var, fill)
}
}
