% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_add_columns.R
\name{add_label_position_typ_1}
\alias{add_label_position_typ_1}
\title{Add centered y-coordinates for filtered data labels of vertical stacked bar
charts}
\usage{
add_label_position_typ_1(df, x_var, y_var, facet_var, filter_cutoff = 0.04)
}
\arguments{
\item{df}{Data frame}

\item{x_var}{Required variable name for the variable containing the discrete
x-coordinates.}

\item{y_var}{Required variable name for the variable containing the continuous
y-coordinates.}

\item{facet_var}{Optional variable name for the discrete variable to facet
by, defaults to NULL.}

\item{filter_cutoff}{Optional cutoff value for the suppression of data
labels. By default, all values below 0.04 of the total value of the
stacked bar chart are suppressed.}
}
\value{
Data frame with additional column "label_" + the name of the
   y-coordinate variable
}
\description{
Calculates the y-coordinates for vertical stacked bar charts, centered for
   each group. The default function does not work if some labels are
   filtered.
}
\examples{
x <- c("WiSe 13/14", "WiSe 13/14", "WiSe 13/14", "WiSe 13/14")
y <- c(1989, 58, 163, 470)
fill <- c("Bachelor 2-Fächer", "Master 1-Fach", "Master 2-Fächer", "Master of Education")
filter_cutoff <- 0.05
df <- add_label_position_typ_1(df, x_var, y_var)
}
