% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_add_columns.R
\name{add_label_position_type_3}
\alias{add_label_position_type_3}
\title{Add centered y-coordinates for filtered data labels of horizontal stacked bar
charts that are scaled to 100\%}
\usage{
add_label_position_type_3(
  df,
  x_var,
  y_var,
  facet_var,
  fill_reverse = FALSE,
  filter_cutoff = 0.04
)
}
\arguments{
\item{df}{Data Frame bzw. Tibble}

\item{x_var}{Required variable name for the variable containing the continuous
x-coordinates.}

\item{y_var}{Required variable name for the variable containing the discrete
y-coordinates.}

\item{facet_var}{Optional variable name for the discrete variable to facet
by, defaults to NULL.}

\item{filter_cutoff}{Optional cutoff value for the suppression of data
labels. By default, all values below 0.04 of the total value of the
stacked bar chart are suppressed.}
}
\value{
Data frame with additional column "label_" + the name of the
   y-coordinate variable
}
\description{
Calculates the y-coordinates for horizontal stacked bar charts, centered for
   each group. The default function does not work if some labels are
   filtered.
}
\examples{
add_label_position_type_3(df, x_var, y_var, fill)
}
