% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_add_columns.R
\name{add_report_author}
\alias{add_report_author}
\title{Adds column \code{report_author} to data frame}
\usage{
add_report_author(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with additional column \code{report_author}
}
\description{
Adds column \code{report_author} to data frame
}
\examples{
\dontrun{
add_report_author(df)
}
}
