% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_add_columns.R
\name{add_report_title}
\alias{add_report_title}
\title{Adds column \code{report_title} to data frame}
\usage{
add_report_title(df)
}
\arguments{
\item{df}{}
}
\value{
Data frame with additional column \code{report_title}
}
\description{
Adds column \code{report_title} to data frame
}
\examples{
add_report_title(df)
}
