% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot.R
\name{add_rub_plot_typ_4}
\alias{add_rub_plot_typ_4}
\title{Helper Function}
\usage{
add_rub_plot_typ_4(df, x, y, group, group_label)
}
\arguments{
\item{df}{das}

\item{x}{das}

\item{y}{das}

\item{group}{das}

\item{group_label}{das}
}
\value{
List of ggplot2 expressions
}
\description{
Helper Function
}
\examples{
add_rub_plot_typ_4(df, x, y, group, group_label)
}
