% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot - Kopie.R, R/rub_plot.R
\name{add_rub_plot_type_4}
\alias{add_rub_plot_type_4}
\title{Helper Function}
\usage{
add_rub_plot_type_4(
  df_t4,
  x_var,
  y_var,
  group_var,
  group_label = NULL,
  base_size = 11,
  base_family = "RubFlama",
  color = RUB_colors["blue"],
  palette_reverse = FALSE
)

add_rub_plot_type_4(
  df_t4,
  x_var,
  y_var,
  group_var,
  group_label = NULL,
  base_size = 11,
  base_family = "RubFlama",
  color = RUB_colors["blue"],
  palette_reverse = FALSE
)
}
\arguments{
\item{df_t4}{Data frame}

\item{x_var}{Required variable name for the variable containing the discrete
x-coordinates.}

\item{y_var}{Required variable name for the variable containing the
continuous y-coordinates.}

\item{group_var}{Variable name for the discrete variable which determines the
groups forming one line, e.g. degree_id.}

\item{group_label}{Optional variable name for the character variable
containing the names of the group variable (e.g. degree_txt), defaults to #'     NULL.}

\item{base_size}{base font size, defaults to 11}

\item{base_family}{base font family, defaults to RubFlama}

\item{color}{Color for font and borders, defaults to `RUB_colors["blue"]`,
i.e. #003560.}

\item{palette_reverse}{Optional boolean indicating whether the colors in the
palette should be reversed, defaults to FALSE.}
}
\value{
List of ggplot2 expressions

List of ggplot2 expressions
}
\description{
Helper Function

Helper Function
}
\examples{
add_rub_plot_type_4(df, x, y, group, group_label)
add_rub_plot_type_4(df, x, y, group, group_label)
}
