% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_bind_rows.R
\name{bind_subject_group}
\alias{bind_subject_group}
\title{Bind rows of subject group data to data frame}
\usage{
bind_subject_group(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with subject group data for each report_nr. This means
   that the subject group data is repeated several times over. Binding the
   rows to each report_nr makes filtering much easier, though.
}
\description{
Bind rows of subject group data to data frame
}
\examples{
\dontrun{
bind_subject_group(df)
}
}
