% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{clean_file_name}
\alias{clean_file_name}
\title{Clean file name of special characters}
\usage{
clean_file_name(file_name)
}
\arguments{
\item{file_name}{}
}
\value{
String
}
\description{
Clean file name of special characters
}
\examples{
clean_file_name("Test_File-Name__X__")
}
