% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_fake_data.R
\name{create_fake_values_type_2}
\alias{create_fake_values_type_2}
\title{Create faked value percentages for figure type 2}
\usage{
create_fake_values_type_2(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with faked values
}
\description{
Create faked value percentages for figure type 2
}
\examples{
\dontrun{
create_fake_values_type_2(df)
}
}
