% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_nrw_213}
\alias{db_nrw_213}
\title{Table series retrieved from...}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1281472 rows and 17 columns.
}
\usage{
db_nrw_213
}
\description{
See vignette for details...
}
\keyword{datasets}
