% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{filter_destatis_code}
\alias{filter_destatis_code}
\title{Filter data frame using the Destatis table code, discarding all columns that
only consist of NA.}
\usage{
filter_destatis_code(df, tablename)
}
\arguments{
\item{df}{Data frame with \code{tablename} column}

\item{tablename}{Destatis table code}
}
\value{
Filtered data frame with all NA-columns dropped
}
\description{
Filter data frame using the Destatis table code, discarding all columns that
only consist of NA.
}
\examples{
\dontrun{
filter_destatis_code(df = db_nrw_213, tablename = "21391KF061")
}
}
