% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data_manipulation.R
\name{filter_label_typ_2}
\alias{filter_label_typ_2}
\title{Filtere Datenlabels anhand eines Schwellwerts}
\usage{
filter_label_typ_2(df, x, y, facet, filter_cutoff = 0.04)
}
\arguments{
\item{df}{Dataframe bzw. Tibble}

\item{x}{Name der Variable, die als x-Koordinate verwendet wird}

\item{y}{Name der Variable, die als y-Koordinate verwendet wird}

\item{facet}{Name der Variable, die die Abbildung in Facets aufteilt}

\item{filter_cutoff}{Schwellwert, ab dem Datenlabels unterdr체ckt werden (z.B. 0.05)}
}
\value{
Gefilterter Dataframe bzw. Tibble
}
\description{
Filtert in gestapelten S채ulendiagrammen die Datenlabels.).
}
\examples{
x <- c("WiSe 13/14", "WiSe 13/14", "WiSe 13/14", "WiSe 13/14")
y <- c(1989, 58, 163, 470)
fill <- c("Bachelor 2-F채cher", "Master 1-Fach", "Master 2-F채cher", "Master of Education")
filter_cutoff <- 0.05
df <- filter_label_y(x, y, fill, y_label, filter_cutoff)
ggplot2::ggplot() +
  ggplot2::geom_bar(aes(x = x, y = y, fill = fill), stat = "identity") +
  ggplot2::geom_label(data = df, aes(x = x, y = y_label, group = fill, label = y))
}
