% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{filter_report}
\alias{filter_report}
\title{Filtere Datensatz auf eine Berichtsnummer plus jeweils relevante Vergleichsgruppe}
\usage{
filter_report(df, report_nr)
}
\arguments{
\item{df}{Dataframe bzw. Tibble mit den Berichtsdaten}

\item{report_nr}{Berichtsnummer}
}
\value{
Dataframe bzw. Tibble mit den gefilterten Berichtsdaten
}
\description{
Filtere Datensatz auf eine Berichtsnummer plus jeweils relevante Vergleichsgruppe
}
\examples{
filter_report(df, report_nr = 12)
}
