% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data_manipulation.R
\name{fix_data_errors}
\alias{fix_data_errors}
\title{Fixes data errors in report numbers 42, 43, 61 and 62.}
\usage{
fix_data_errors(df)
}
\arguments{
\item{df}{}
}
\value{
Data frame with fixes to the columns \code{faculty_txt},
   \code{report_nr}, and \code{figure_sort}
}
\description{
Fixes data errors in report numbers 42, 43, 61 and 62.
}
\examples{
fix_data_errors(df)
}
