% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data_manipulation.R
\name{generate_cohort_data_2018_FG}
\alias{generate_cohort_data_2018_FG}
\title{Generate cohort data on the level of subject degree groups}
\usage{
generate_cohort_data_2018_FG(df_cohort)
}
\arguments{
\item{df_cohort}{Data frame}
}
\value{
Data frame with cohort data for report_id "2018_FG"
}
\description{
Generate cohort data on the level of subject degree groups
}
\examples{
\dontrun{
generate_cohort_data_2018_FG(df_cohort)
}
}
