% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{generate_orga}
\alias{generate_orga}
\title{Erzeuge Übersicht aller Berichte aus den Daten}
\usage{
generate_orga(df)
}
\arguments{
\item{df}{Dataframe bzw. Tibble mit den Berichtsdaten}
}
\value{
Dataframe bzw. Tibble mit Berichtsübersicht
}
\description{
Erzeuge Übersicht aller Berichte aus den Daten
}
\examples{
\dontrun{
generate_orga(df)
}
}
