% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_bind_rows.R,
%   R/wrangle_data_manipulation.R
\name{get_2018_fg}
\alias{get_2018_fg}
\title{Get data frame with generated data for \code{report_id} "2018_FG"}
\usage{
get_2018_fg(df)

get_2018_fg(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with all data for \code{report_id} "2018_FG"

Data frame with all data for \code{report_id} "2018_FG"
}
\description{
Get data frame with generated data for \code{report_id} "2018_FG"

Get data frame with generated data for \code{report_id} "2018_FG"
}
\examples{
get_2018_fg(df)
get_2018_fg(df)
}
