% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_author}
\alias{get_author}
\title{Get author for report title page}
\usage{
get_author(param_list)
}
\arguments{
\item{df}{}
}
\value{
String
}
\description{
Get author for report title page
}
\examples{
get_author(df)
}
