% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data_manipulation.R
\name{get_cohort_data}
\alias{get_cohort_data}
\title{Get cohort data, including generated cohort data for report_id 2018_FG}
\usage{
get_cohort_data(df_orga)
}
\arguments{
\item{df_orga}{Data frame with orga}
}
\value{
Data frame with corrected cohort analysis data
}
\description{
Get cohort data, including generated cohort data for report_id 2018_FG
}
\examples{
\dontrun{
get_cohort_data(df_orga)
}
}
