% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_fake_data.R
\name{get_fake_data_type_1_4}
\alias{get_fake_data_type_1_4}
\title{Get data frame with fake values for figure types 1 and 4}
\usage{
get_fake_data_type_1_4(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with fake values for value_n_total
}
\description{
Get data frame with fake values for figure types 1 and 4
}
\examples{
\dontrun{
get_fake_data_type_1_4(df)
}
}
