% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_fake_data.R
\name{get_fake_data_type_2}
\alias{get_fake_data_type_2}
\title{Create fake values for figure type 2}
\usage{
get_fake_data_type_2(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with fake values for figure type 2
}
\description{
Create fake values for figure type 2
}
\examples{
\dontrun{
get_fake_data_type_2(df)
}
}
