% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_fake_data.R
\name{get_fake_data_type_3}
\alias{get_fake_data_type_3}
\title{Create fake values for figure type 3}
\usage{
get_fake_data_type_3(df, report_id_group_level = NULL)
}
\arguments{
\item{df}{Data frame}

\item{report_id_group_level}{A vector of report_ids, indicating which reports
do not have hierarchical reference group comparisons}
}
\value{
Data frame with fake values for figure type 3
}
\description{
Create fake values for figure type 3
}
\examples{
\dontrun{
get_fake_data_type_3(df, report_id_group_level = c(62, 64))
}
}
