% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_fake_data.R
\name{get_fake_df}
\alias{get_fake_df}
\title{Get data frame with generated fake values}
\usage{
get_fake_df(df, report_id_group_level = NULL)
}
\arguments{
\item{df}{Data frame}

\item{report_id_group_level}{A vector of report_ids, indicating which reports
do not have reference group comparisons plotted for each figure.}
}
\value{
Data frame with fake values for \code{axis_label},
   \code{value_n_total} and \code{value_percentage}.
}
\description{
Get data frame with generated fake values
}
\examples{
\dontrun{
get_fake_df(df, report_id_group_level = c(62, 64))
}
}
