% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_fake_data.R
\name{get_fake_values_type_1_4}
\alias{get_fake_values_type_1_4}
\title{Get vector with faked values for figure types 1 and 4}
\usage{
get_fake_values_type_1_4(figure_type_id, length, base_value = NULL)
}
\arguments{
\item{figure_type_id}{Integer}

\item{length}{Integer}

\item{base_value}{Integer}
}
\value{
Vector with fake values
}
\description{
Get vector with faked values for figure types 1 and 4
}
\examples{
get_fake_values_type_1_4(
   figure_type_id = 4,
   length = 5,
   base_value = 800
   )
}
