% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_fake_data.R
\name{get_fake_values_type_2}
\alias{get_fake_values_type_2}
\title{Get vector with faked value percentages for figure type 2}
\usage{
get_fake_values_type_2(fractions = 5, periods = 8)
}
\arguments{
\item{fractions}{Number of fractions}

\item{periods}{Number of time periods}
}
\value{
Vector with faked value percentages
}
\description{
Get vector with faked value percentages for figure type 2
}
\examples{
get_fake_values_type_2(
   fractions = 6,
   periods = 3
   )
}
