% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_fake_data.R
\name{get_fake_values_type_3}
\alias{get_fake_values_type_3}
\title{Get faked value percentages for figure type 3}
\usage{
get_fake_values_type_3(x)
}
\arguments{
\item{x}{Length}
}
\value{

}
\description{
Get faked value percentages for figure type 3
}
\examples{
get_fake_values_type_3(5)
}
