% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{get_fallback_font_df}
\alias{get_fallback_font_df}
\title{Get system dependent fallback font if a given font is not available}
\usage{
get_fallback_font_df(
  fonts = systemfonts::system_fonts(),
  fallback_alias = "sans"
)
}
\arguments{
\item{fonts}{Data frame obtained by \code{systemfonts::system_fonts}}

\item{fallback_alias}{Character, defaults to "sans"}
}
\value{
Data frame obtained by \code{systemfonts::system_fonts} with one row
}
\description{
Get system dependent fallback font if a given font is not available
}
\examples{
get_fallback_font_df()
}
