% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_get_functions.R
\name{get_figure_caption}
\alias{get_figure_caption}
\title{Get figure caption}
\usage{
get_figure_caption(param_list)
}
\arguments{
\item{param_list}{List of parameters containing figure_type_id, figure_id,
degree_group_txt, subject_area_txt, subject_group_txt, report_id}
}
\value{
String containing figure caption
}
\description{
Get figure caption
}
\examples{
\dontrun{
get_figure_caption(param_list)
}
}
