% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_template.R
\name{get_figure_chunk_df}
\alias{get_figure_chunk_df}
\title{Get code chunks for plotting all figures}
\usage{
get_figure_chunk_df(
  df,
  font_family = get_font_df()[["family"]],
  function_call = "RUBer::plot_figure"
)
}
\arguments{
\item{df}{Data frame with columns \code{figure_count}}

\item{font_family}{Character, the font family to use for all plots, defaults
to \code{get_font_df()[["family"]]}}

\item{function_call}{Character, the function call for each plot chunk,
defaults to \code{RUBer::plot_figure}}
}
\value{
Data frame with columns \code{figure_count}, \code{chunk_heading},
\code{chunk_subheading}, \code{chunk_figure_df} and \code{chunk_plot_figure}
}
\description{
Get code chunks for plotting all figures
}
\examples{
get_figure_chunk_df(RUBer::df_example \%>\% dplyr::filter(report_nr == 1L))
}
