% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_get_functions.R
\name{get_figure_df}
\alias{get_figure_df}
\title{Prepare data frame for plotting of a particular figure by filtering and
   setting factors.}
\usage{
get_figure_df(df, figure_count)
}
\arguments{
\item{df}{Data frame}

\item{figure_count}{Figure count of the figure}
}
\value{
Data frame
}
\description{
Prepare data frame for plotting of a particular figure by filtering and
   setting factors.
}
\examples{
get_figure_df(df_fake, figure_count = 16)
}
