% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_get_functions.R
\name{get_figure_height}
\alias{get_figure_height}
\title{Get figure height}
\usage{
get_figure_height(param_list, lower_bound = 2, upper_bound = 9.6)
}
\arguments{
\item{param_list}{Data frame}

\item{lower_bound}{Minimum height (default = 2.25)}

\item{upper_bound}{Maximum height (default = 9.6)}
}
\value{
Figure height in inches
}
\description{
Get figure height
}
\examples{
get_figure_height(param_list)
}
