% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_file_path}
\alias{get_file_path}
\title{Get file path for automatic report generation}
\usage{
get_file_path(file_directory, file_name)
}
\arguments{
\item{file_directory}{String directory}

\item{file_name}{Required string containing file name with file extension}
}
\value{
String file path
}
\description{
Get file path for automatic report generation
}
\examples{
get_file_path(file_name = "test")
}
