% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{get_font_df}
\alias{get_font_df}
\title{Get data frame of font information for a font file}
\usage{
get_font_df(font_file = "RubFlama-Regular.ttf")
}
\arguments{
\item{font_file}{Character, name of the font file (case insensitive)}
}
\value{
Data frame obtained by \code{systemfonts::system_fonts} with one row
}
\description{
Get data frame of font information for a font file
}
\examples{
get_font_df("RUB Scala TZ.ttf")
}
