% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_headings}
\alias{get_headings}
\title{Add headings to report data frame}
\usage{
get_headings(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with four additional columns: "heading", "subheading",
"is_heading" and "is_subheading".
}
\description{
Add headings to report data frame
}
\examples{
\dontrun{
get_headings(df)
}
}
