% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot.R
\name{get_legend_columns}
\alias{get_legend_columns}
\title{Gets number of legend columns based on the plot width and the text base size}
\usage{
get_legend_columns(
  legend_text,
  legend_key_width = plot_width/100,
  legend_key_spacing = plot_width/100,
  plot_width = 6.8,
  base_size = 11
)
}
\arguments{
\item{legend_text}{Vector with the legend text}

\item{legend_key_width}{Legend key width}

\item{legend_key_spacing}{Legend key spacing}

\item{plot_width}{Width of the plot in inches}

\item{base_size}{Text base size (in points)}
}
\value{
Numeric with the number of columns for the legend
}
\description{
Gets number of legend columns based on the plot width and the text base size
}
\examples{
get_legend_columns(
  legend_text = c(
    "1 - eigener Verdienst",
    "2 - Mittel der Eltern/Verwandten",
    "3 - Förderung nach BAföG",
    "4 - Stipendium",
    "5 - Sonstiges"
  )
)
}
