% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data_manipulation.R
\name{get_map_parameters}
\alias{get_map_parameters}
\title{Get function calls and parameters for use of purrr::map}
\usage{
get_map_parameters(df)
}
\arguments{
\item{df}{Dataframe bzw. Tibble}
}
\value{
Dataframe bzw. Tibble with chunk_label and function_param columns
}
\description{
Get function calls and parameters for use of purrr::map
}
\examples{
get_map_paramters(df)
}
