% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_report_nr_by_id}
\alias{get_report_nr_by_id}
\title{Get all unique values of \code{report_nr} for a \code{report_type_id}}
\usage{
get_report_nr_by_id(df, report_type_id)
}
\arguments{
\item{df}{Data frame with columns \code{report_nr} and \code{report_type_id}}

\item{report_type_id}{Any of \code{c("STG", "M_ED", "MED", "FGR", "SZMA")}}
}
\value{
Integer vector
}
\description{
Get all unique values of \code{report_nr} for a \code{report_type_id}
}
\examples{
get_report_nr_by_id(df = df_example, report_type_id = "FGR")
}
