% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_colors.R
\name{get_RUB_colors}
\alias{get_RUB_colors}
\title{Function to extract RUB palette colors as hex codes}
\usage{
get_RUB_colors(...)
}
\arguments{
\item{...}{Character names of RUB palette colors}
}
\value{
Vector with specified color hex codes
}
\description{
Function to extract RUB palette colors as hex codes
}
\examples{
get_RUB_colors("green", "blue")
get_RUB_colors(1, 3, 1)
}
