% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_bind_rows.R
\name{get_subject_group_rhs}
\alias{get_subject_group_rhs}
\title{Get right hand side of generated subject group data}
\usage{
get_subject_group_rhs(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with right hand side of generated subject group data
}
\description{
Get right hand side of generated subject group data
}
\examples{
\dontrun{
get_subject_group_rhs(df)
}
}
