% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_unique}
\alias{get_unique}
\title{Returns all unique values for a specified column in a data frame}
\usage{
get_unique(df, column)
}
\arguments{
\item{df}{Data frame}

\item{column}{Column name}
}
\value{
Vector with unique values
}
\description{
Returns all unique values for a specified column in a data frame
}
\examples{
get_unique(df, report_nr)
}
