% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_bind_rows.R
\name{join_subject_group}
\alias{join_subject_group}
\title{Joins left hand and right hand side of subject group data}
\usage{
join_subject_group(df_lhs, df_rhs)
}
\arguments{
\item{df_lhs}{Left hand side of subject group data}

\item{df_rhs}{Right hand side of subject group data}
}
\value{
Joined data frame
}
\description{
Joins left hand and right hand side of subject group data
}
\examples{
join_subject_group(df_lhs, df_rhs)
}
