% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data_manipulation.R
\name{mutate_plot_variables}
\alias{mutate_plot_variables}
\title{Generiert f端r die Abbildungstypen 1-4 Variablen f端r die Ploterstellung}
\usage{
mutate_plot_variables(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with additional columns \code{x}, \code{y}, \code{fill},
    \code{fill_label}, \code{fill_reverse}, \code{facet} und \code{group}.
}
\description{
Generiert f端r die Abbildungstypen 1-4 Variablen f端r die Ploterstellung
}
\examples{
\dontrun{
mutate_plot_variables(df_filtered)
}
}
