% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot - Kopie.R, R/rub_plot.R
\name{plot_discrete_palette}
\alias{plot_discrete_palette}
\title{Get discrete palette for the plot}
\usage{
plot_discrete_palette(colors_n)

plot_discrete_palette(colors_n)
}
\arguments{
\item{colors_n}{Integer for the number of requestes colors in the discrete
palette.}
}
\value{
Name of the appropriate discrete palette from RUB_palettes

Name of the appropriate discrete palette from RUB_palettes
}
\description{
At the moment, RUB_palettes has discrete palettes for up to eight unique
    colors. Above that number, colors are interpolated.

At the moment, RUB_palettes has discrete palettes for up to eight unique
    colors. Above that number, colors are interpolated.
}
\examples{
plot_discrete_palette(12)
plot_discrete_palette(12)
}
