% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot - Kopie.R, R/rub_plot.R
\name{plot_figure}
\alias{plot_figure}
\title{Plot RUB figure}
\usage{
plot_figure(df)

plot_figure(df)
}
\arguments{
\item{df}{Data Frame bzw. Tibble}
}
\value{
ggplot object

ggplot object
}
\description{
Plot RUB figure

Plot RUB figure
}
\examples{
\dontrun{
plot_figure(df)
}
\dontrun{
plot_figure(df)
}
}
