% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{post_process}
\alias{post_process}
\title{Post process Word file created with officedown}
\usage{
post_process(old_path, new_path, overwrite = FALSE)
}
\arguments{
\item{old_path}{String with path to Word file created by officedown}

\item{new_path}{New path for edited Word file}

\item{overwrite}{Boolean, overwrite files if they exist. If this is \code{FALSE}
and the file exists an error will be thrown.}
}
\value{
side effects
}
\description{
Post process Word file created with officedown
}
\keyword{internal}
