% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data_manipulation.R
\name{read_cohort_data}
\alias{read_cohort_data}
\title{Reads csv file with cohort data}
\usage{
read_cohort_data()
}
\value{
Data frame with cohort data
}
\description{
Reads csv file with cohort data
}
\examples{
\dontrun{
read_cohort_data()
}
}
