% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{register_font}
\alias{register_font}
\title{Register font contained in font data frame}
\usage{
register_font(font_df = RUBer::get_font_df())
}
\arguments{
\item{font_df}{Data frame with one row obtained by \code{RUBer::get_font_df()}}
}
\value{
Side effects
}
\description{
Register font contained in font data frame
}
\examples{
register_font()
}
