% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{register_font_flama}
\alias{register_font_flama}
\title{Registers RUB Flama font to be used with \code{showtext} package}
\usage{
register_font_flama(
  family = get_font_df()[["family"]],
  regular = get_font_df()[["path"]],
  bold = get_font_df("RubFlama-Bold.ttf")[["path"]],
  italic = get_font_df("RubFlama-Italic.ttf")[["path"]],
  bolditalic = get_font_df("RubFlama-BoldItalic.ttf")[["path"]]
)
}
\arguments{
\item{family}{Character, font family, defaults to
\code{get_font_df()[["family"]]}}

\item{regular}{Character, path of the font file for "regular" font style
defaults to \code{get_font_df()[["path"]]}}

\item{bold}{Character, path of the font file for "bold" font style,
defaults to \code{get_font_df("RubFlama-Bold.ttf")[["path"]]}}

\item{italic}{Character, path of the font file for "italic" font style,
\code{get_font_df("RubFlama-Italic.ttf")[["path"]]}}

\item{bolditalic}{Character, path of the font file for "bold italic" font style,
defaults to \code{get_font_df("RubFlama-BoldItalic.ttf")[["path"]]}}
}
\value{

}
\description{
Registers RUB Flama font to be used with \code{showtext} package
}
\examples{
register_font_flama()
}
