% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{render_report}
\alias{render_report}
\title{Render report}
\usage{
render_report(
  p_df,
  report_nr,
  rmd_template = here::here("datenreport_new.Rmd"),
  date = format(Sys.Date(), format = "\%B \%Y")
)
}
\arguments{
\item{p_df}{Data frame containing the data for all reports}

\item{report_nr}{Report number of the report}

\item{rmd_template}{Path to the R Markdown File for that report, defaults to
\code{here::here("datenreport_new.Rmd")}}

\item{date}{Date of the report displayed on the title page, defaults to
\code{format(Sys.Date(), format= "\%B \%Y")}.}
}
\description{
Render report
}
\examples{
\dontrun{
render_report(df, 12)
}
}
