% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{replace_first_page_header}
\alias{replace_first_page_header}
\title{Create header with option "different first page" enabled}
\usage{
replace_first_page_header(doc)
}
\arguments{
\item{doc}{an rdocx object}
}
\value{
A modified rdocx object
}
\description{
Create header with option "different first page" enabled
}
\keyword{internal}
