% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{replace_toc}
\alias{replace_toc}
\title{Replace level of TOC field}
\usage{
replace_toc(doc, style, new_level)
}
\arguments{
\item{doc}{an rdocx object}

\item{style}{text, the style referenced by the TOC field}

\item{new_level}{numeric, the new level for the TOC field}
}
\value{
A modified rdocx object
}
\description{
Replace level of TOC field
}
\keyword{internal}
