% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_colors.R
\docType{data}
\name{RUB_colors}
\alias{RUB_colors}
\title{RUB colors}
\format{An object of class \code{character} of length 19.}
\usage{
RUB_colors
}
\description{
These are available RUB colors as defined by the corporate design.
}
\references{
https://www.ruhr-uni-bochum.de/cd/
}
\keyword{datasets}
