% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rub_format_percent}
\alias{rub_format_percent}
\title{Format flextable columns with percentages (percentages, NAs)}
\usage{
rub_format_percent(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Formatted vector
}
\description{
Format flextable columns with percentages (percentages, NAs)
}
\examples{
x <- c("0.29", NA)
rub_format_percent(x)
}
