% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_colors.R
\docType{data}
\name{RUB_palettes}
\alias{RUB_palettes}
\title{RUB color palette}
\format{An object of class \code{list} of length 12.}
\usage{
RUB_palettes
}
\description{
This is a color palette using the colors defined by the corporate design.
}
\references{
https://www.ruhr-uni-bochum.de/cd/
}
\keyword{datasets}
