% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot.R
\name{rub_plot_typ_1}
\alias{rub_plot_typ_1}
\title{Plot vertical stacked bar chart (figure type 1)}
\usage{
rub_plot_typ_1(df, x, y, y_label = "", fill, fill_label = fill,
  caption = "", filter_cutoff = 0.04, facet_var = NULL)
}
\arguments{
\item{df}{Data frame}

\item{x}{Required variable name for the variable containing the discrete
x-coordinates}

\item{y}{Required variable name for the variable containing the continuous
y-coordinates}

\item{y_label}{Optional label for the y-axis, defaults to an empty string.}

\item{fill}{Variable name for the discrete variable which determines the
groups to be stacked, e.g. degree}

\item{fill_label}{Optional variable name for the character variable
containing the names of the fill variable, defaults to fill.}

\item{caption}{Optional character containing the data source for the figure
(prefix 'Quelle:' is automatically added)}

\item{filter_cutoff}{Optional cutoff value for the suppression of data
labels. By default, all values below 0.04 of the total value of the
stacked bar chart are suppressed.}

\item{facet_var}{Optional variable name to facet by, defaults to NULL.}
}
\value{
A ggplot object
}
\description{
vertical stacked bar chart in the RUB corporate design. The variables x, y
    and fill are mandatory, all others are optional.
}
\examples{
df_type_1 <- tibble::tribble(
~x, ~y, ~fill,
"WiSe 13/14", 120, "Bachelor 1-Fach",
"WiSe 14/15", 105, "Bachelor 1-Fach",
"WiSe 15/16", 124, "Bachelor 1-Fach",
"WiSe 16/17", 114, "Bachelor 1-Fach",
"WiSe 17/18", 122, "Bachelor 1-Fach",
"WiSe 13/14", 121, "Master 1-Fach",
"WiSe 14/15", 129, "Master 1-Fach",
"WiSe 15/16", 122, "Master 1-Fach",
"WiSe 16/17", 168, "Master 1-Fach",
"WiSe 17/18", 7, "Master 1-Fach",
)

rub_plot_typ_1(
  df = df_type_1,
  x = x,
  y = y,
  fill = fill
)
}
