% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot.R
\name{rub_plot_typ_1_and_4}
\alias{rub_plot_typ_1_and_4}
\title{Zeichne gruppiertes Liniendiagramm (Abbildungstyp 5)}
\usage{
rub_plot_typ_1_and_4(df)
}
\arguments{
\item{df}{Data Frame bzw. Tibble}
}
\value{
Ein ggplot Objekt
}
\description{
Zeichne gruppiertes Liniendiagramm (Abbildungstyp 5)
}
\examples{
rub_plot_typ_1_and_4(df = df)
}
