% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot.R
\name{rub_plot_typ_4}
\alias{rub_plot_typ_4}
\title{Zeichne gruppiertes Liniendiagramm (Abbildungstyp 4)}
\usage{
rub_plot_typ_4(df, x, y, group, group_label, caption = "")
}
\arguments{
\item{df}{Data Frame bzw. Tibble}

\item{x}{x-Position der Linienpunkte}

\item{y}{y-Position der Linienpunkte}

\item{group}{Gruppierungsvariable f체r die einzelnen Linien (z.B. Abschluss ID)}

\item{group_label}{Beschriftung der Gruppierungsvariable (z.B. Abschluss DTXT)}

\item{caption}{Quellenangabe f체r die Abbildung (Pr채fix 'Quelle:' wird automatisch hinzugef체gt)}
}
\value{
Ein ggplot Objekt
}
\description{
Zeichne gruppiertes Liniendiagramm (Abbildungstyp 4)
}
\examples{
rub_plot_typ_4(df = df_fig_t4, x = time, y = value_n_total, group = degree_sort, group_label = degree_txt )
}
