% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot.R
\name{rub_plot_type_1}
\alias{rub_plot_type_1}
\title{Plot vertical stacked bar chart (figure type 1)}
\usage{
rub_plot_type_1(
  df,
  x_var,
  x_var_label = NULL,
  y_var,
  y_axis_label = "",
  fill_var,
  fill_reverse = FALSE,
  fill_label = NULL,
  caption = "",
  caption_prefix = "Quelle:",
  filter_cutoff = 0.04,
  facet_var = NULL,
  color = RUB_colors["blue"],
  palette_reverse = FALSE,
  base_family = get_font_df()[["family"]],
  base_size = 11,
  plot_width = 6.8
)
}
\arguments{
\item{df}{Data frame}

\item{x_var}{Required variable name for the variable containing the discrete
x-coordinates.}

\item{x_var_label}{Optional variable name for the character variable
containing the names of the x variable, defaults to NULL.}

\item{y_var}{Required variable name for the variable containing the
y-coordinates. Will be coerced to numeric with \code{as.numeric}.}

\item{y_axis_label}{Optional label for the y-axis, defaults to an empty
string.}

\item{fill_var}{Variable name for the discrete variable which determines the
groups to be stacked, e.g. degree.}

\item{fill_reverse}{Boolean indicating whether the order of the fill variable
should be reversed, defaults to FALSE.}

\item{fill_label}{Optional variable name for the character variable
containing the names of the fill variable, defaults to NULL.}

\item{caption}{Optional character containing the data source for the figure
(prefix 'Quelle:' is automatically added).}

\item{caption_prefix}{Optional character containing the prefix for the
caption, defaults to 'Quelle:'.}

\item{filter_cutoff}{Optional cutoff value for the suppression of data
labels. By default, all values below 0.04 of the total value of the
stacked bar chart are suppressed.}

\item{facet_var}{Optional variable name for the discrete variable to facet
by, defaults to NULL.}

\item{color}{Color for font and borders, defaults to \code{RUB_colors["blue"]},
i.e. #003560.}

\item{palette_reverse}{Optional boolean indicating whether the colors in the
palette should be reversed, defaults to FALSE.}

\item{base_family}{base font family, defaults to RubFlama}

\item{base_size}{base font size, defaults to 11}

\item{plot_width}{Width of the plot in inches, defaults to 6.8}
}
\value{
A ggplot object
}
\description{
Vertical stacked bar chart in the RUB corporate design. The variables x_var,
y_var and fill_var are required, all others are optional.
}
\examples{
# Create test values for all three mandatory variables (x_var, y_var, fill_var).
df_t1_ex1 <- tibble::tribble(
  ~term, ~students, ~degree,
  "Spring '13", 120, "Bachelor 1-Subject",
  "Spring '14", 105, "Bachelor 1-Subject",
  "Spring '15", 124, "Bachelor 1-Subject",
  "Spring '16", 114, "Bachelor 1-Subject",
  "Spring '17", 122, "Bachelor 1-Subject",
  "Spring '13", 121, "Master 1-Subject",
  "Spring '14", 129, "Master 1-Subject",
  "Spring '15", 122, "Master 1-Subject",
  "Spring '16", 168, "Master 1-Subject",
  "Spring '17", 7, "Master 1-Subject",
)

# x_var is mapped to term, y_var to students, and the fill_var to degree.
# base_size increases the text sizes from the default, 11, to 14. The font
# family is changed from "RubFlama" to "sans" (available on all systems).
rub_plot_type_1(
  df = df_t1_ex1,
  x_var = term,
  y_var = students,
  fill_var = degree,
  base_size = 14,
  base_family = "sans"
)
}
\seealso{
Other rub_plot_types: 
\code{\link{rub_plot_type_1_and_4}()},
\code{\link{rub_plot_type_2}()},
\code{\link{rub_plot_type_3}()},
\code{\link{rub_plot_type_4}()}
}
\concept{rub_plot_types}
