% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot.R
\name{rub_plot_type_1_and_4}
\alias{rub_plot_type_1_and_4}
\title{Plot grouped line chart on top of vertical stacked bar chart (combination chart of figure types 1 and 4)}
\usage{
rub_plot_type_1_and_4(
  df,
  x_var,
  x_axis_label = "",
  y_var,
  y_axis_label = "",
  fill_var,
  fill_reverse = FALSE,
  fill_label = NULL,
  group_var,
  group_label = NULL,
  caption = "",
  caption_prefix = "Quelle:",
  filter_cutoff = 0.04,
  facet_var = NULL,
  color = RUB_colors["blue"],
  palette_reverse = FALSE,
  base_family = "RubFlama",
  base_size = 11,
  plot_width = 6.8
)
}
\arguments{
\item{df}{Data frame}

\item{x_var}{Required variable name for the variable containing the discrete
x-coordinates.}

\item{x_axis_label}{Optional label for the x-axis, defaults to an empty
string.}

\item{y_var}{Required variable name for the variable containing the
continuous y-coordinates.}

\item{y_axis_label}{Optional label for the y-axis, defaults to an empty
string.}

\item{fill_var}{Variable name for the discrete variable which determines the
groups to be stacked, e.g. degree.}

\item{fill_reverse}{Boolean indicating whether the order of the fill variable
should be reversed, defaults to FALSE.}

\item{fill_label}{Optional variable name for the character variable
containing the names of the fill variable, defaults to NULL.}

\item{group_var}{Variable name for the discrete variable which determines the
groups forming one line, e.g. degree_id.}

\item{group_label}{Optional variable name for the character variable
containing the names of the group variable (e.g. degree_txt), defaults to #'     NULL.}

\item{caption}{Optional character containing the data source for the figure
(prefix 'Quelle:' is automatically added).}

\item{caption_prefix}{Optional character containing the prefix for the
caption, defaults to 'Quelle:'.}

\item{filter_cutoff}{Optional cutoff value for the suppression of data
labels. By default, all values below 0.04 of the total value of the
stacked bar chart are suppressed.}

\item{facet_var}{Optional variable name for the discrete variable to facet
by, defaults to NULL.}

\item{color}{Color for font and borders, defaults to \code{RUB_colors["blue"]},
i.e. #003560.}

\item{palette_reverse}{Optional boolean indicating whether the colors in the
palette should be reversed, defaults to FALSE.}

\item{base_family}{base font family, defaults to RubFlama}

\item{base_size}{base font size, defaults to 11}

\item{plot_width}{Width of the plot in inches, defaults to 6.8}
}
\value{
A ggplot object
}
\description{
Plot grouped line chart on top of vertical stacked bar chart (combination chart of figure types 1 and 4)
}
\examples{
\dontrun{
rub_plot_type_1_and_4(df = df)
}
}
