% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot.R
\name{rub_plot_type_3}
\alias{rub_plot_type_3}
\title{Plot horizontal stacked bar charts that are scaled to 100\% (figure type 3)}
\usage{
rub_plot_type_3(
  df,
  x_var,
  y_var,
  y_axis_label = "",
  fill_var,
  fill_label = NULL,
  fill_reverse = FALSE,
  facet_var = NULL,
  group_var = NULL,
  caption = "",
  caption_prefix = "Quelle:",
  filter_cutoff = 0.04,
  color = RUB_colors["blue"],
  base_family = "RubFlama",
  base_size = 11
)
}
\arguments{
\item{df}{Data Frame bzw. Tibble}

\item{x_var}{Required variable name for the variable containing the continuous
x-coordinates.}

\item{y_var}{Required variable name for the variable containing the discrete
y-coordinates.}

\item{y_axis_label}{Optional label for the y-axis, defaults to an empty
string.}

\item{fill_var}{Variable name for the discrete variable which determines the
groups to be stacked, e.g. degree.}

\item{fill_label}{Optional variable name for the character variable
containing the names of the fill variable, defaults to NULL.}

\item{fill_reverse}{Boolean indicating whether the order of the fill variable
should be reversed, default = FALSE.}

\item{facet_var}{Optional variable name for the discrete variable to facet
by, defaults to NULL.}

\item{caption}{Optional character containing the data source for the figure
(prefix 'Quelle:' is automatically added).}

\item{caption_prefix}{Optional character containing the prefix for the
caption, defaults to 'Quelle:'.}

\item{filter_cutoff}{Optional cutoff value for the suppression of data
labels. By default, all values below 0.04 of the total value of the
stacked bar chart are suppressed.}

\item{color}{Color for font and borders, defaults to `RUB_colors["blue"]`,
i.e. #003560.}

\item{base_family}{base font family, defaults to RubFlama}

\item{base_size}{base font size, defaults to 11}
}
\value{
A ggplot object
}
\description{
Plot horizontal stacked bar charts that are scaled to 100\% (figure type 3)
}
\examples{
# Create test values for all three mandatory variables (x_var, y_var, fill_var)
df_t3_ex1 <- tibble::tribble(
 ~survey_group, ~item_value, ~item_value_percentage,
 "Bachelor 1-Subject (n=400)", "Exceeded prescribed period of study", 0.3070448,
 "Bachelor 1-Subject (n=400)", "Within prescribed period of study", 0.6929552,
 "SG Bachelor 1-Subject (n=669)", "Exceeded prescribed period of study", 0.1122486,
 "SG Bachelor 1-Subject (n=669)", "Within prescribed period of study", 0.8877514
)

rub_plot_type_3(
 df = df_t3_ex1,
 x_var = item_value_percentage,
 y_var = survey_group,
 fill_var = item_value,
)
}
