% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_tables.R
\name{rub_table_eb}
\alias{rub_table_eb}
\title{Get formatted flextable of response rates for the Eingangsbefragung}
\usage{
rub_table_eb(df, typology, headings)
}
\arguments{
\item{df}{Data frame}

\item{typology}{Data frame with flextable typology}

\item{headings}{Character vectors of headings}
}
\value{
Formatted Flextable
}
\description{
Get formatted flextable of response rates for the Eingangsbefragung
}
