% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_tables.R
\name{rub_table_item}
\alias{rub_table_item}
\title{Get formatted flextable of items with largest deviation from comparison group}
\usage{
rub_table_item(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Formatted Flextable
}
\description{
Get formatted flextable of items with largest deviation from comparison group
}
