% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_tables.R
\name{rub_table_metrics}
\alias{rub_table_metrics}
\title{Get formatted flextable of metrics}
\usage{
rub_table_metrics(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Formatted flextable
}
\description{
Get formatted flextable of metrics
}
