% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_tables.R
\name{rub_table_programs}
\alias{rub_table_programs}
\title{Get formatted flextable of funded projects}
\usage{
rub_table_programs(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Formatted Flextable
}
\description{
Get formatted flextable of funded projects
}
