% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_tables.R
\name{rub_table_vb}
\alias{rub_table_vb}
\title{Get formatted flextable of response rates for the Verlaufsbefragung}
\usage{
rub_table_vb(df, typology, headings)
}
\arguments{
\item{df}{Data frame}

\item{typology}{Data frame with flextable typology}

\item{headings}{Character vectors of headings}
}
\value{
Formatted Flextable
}
\description{
Get formatted flextable of response rates for the Verlaufsbefragung
}
