% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUBer-package.R
\docType{package}
\name{RUBer-package}
\alias{RUBer}
\alias{RUBer-package}
\title{RUBer: Parameterized Microsoft Word reports in the RUB corporate design}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

RUBer is an R package created for parameterized reporting
    according to the requirements at the Ruhr-Universität Bochum (RUB).
    The package provides a large example dataset containing fake values,
    an R Markdown template in the RUB corporate design, preconfigured
    ggplot2 plotting functions, custom themes for ggplot2 and flextable,
    as well as functions making the RUB corporate design colors available.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RichardMeyer-Eppler/RUBer
    https://richardmeyer-eppler.github.io/RUBer/}
  \item Report bugs at \url{https://github.com/RichardMeyer-Eppler/RUBer/issues}
}

}
\author{
\strong{Maintainer}: Richard Meyer-Eppler \email{richard.meyer-eppler@uv.ruhr-uni-bochum.de}

}
\keyword{internal}
