% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUBer.R
\docType{package}
\name{Ruber}
\alias{Ruber}
\title{RUBer: Parameterized Microsoft Word reports in the RUB corporate
design}
\description{
RUBer is an R package created for parameterized
reporting according to the requirements at the Ruhr-Universität Bochum
(RUB). The package provides a large example dataset containing fake
values, an R Markdown template in the RUB corporate design,
preconfigured ggplot2 plotting functions, custom themes for ggplot2
and flextable, as well as functions making the RUB corporate design
colors available.
}
\section{RUBer functions}{

The foo functions ...
}

