% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{safely_render_report}
\alias{safely_render_report}
\title{Wrap render_report in purrr::safely when package is loaded}
\usage{
safely_render_report(...)
}
\description{
Wrap render_report in purrr::safely when package is loaded
}
