% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_colors.R
\name{scale_color_rub}
\alias{scale_color_rub}
\title{Color scale constructor for RUB colors}
\usage{
scale_color_rub(palette = "discrete", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in \code{RUB_palettes}}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or
not, defaults to TRUE.}

\item{reverse}{Optional boolean indicating whether the palette should be
reversed, defaults to FALSE.}

\item{...}{Additional arguments passed to
\code{\link[ggplot2]{discrete_scale}} or
\code{\link[ggplot2]{scale_gradient}}, used when discrete is TRUE
or FALSE, respectively}
}
\description{
Color scale constructor for RUB colors
}
\examples{
ggplot2::ggplot(
  data = mtcars,
  ggplot2::aes(
    x = mpg,
    y = disp,
    color = as.factor(carb)
    )
  ) +
  ggplot2::geom_point() +
  scale_color_rub(
    palette = "discrete"
  )
}
