% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_colors.R
\name{scale_fill_rub}
\alias{scale_fill_rub}
\title{Fill scale constructor for RUB colors}
\usage{
scale_fill_rub(palette = "discrete", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in RUB_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to
\code{\link[ggplot2]{discrete_scale}} or
\code{\link[ggplot2]{scale_gradient}}, used when discrete is TRUE
or FALSE, respectively}
}
\description{
Fill scale constructor for RUB colors
}
\examples{
ggplot(mpg, aes(x = trans, y = displ, fill = class)) +
geom_bar(stat = "identity") +
scale_fill_rub()
}
