% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data_manipulation.R
\name{set_factor_var}
\alias{set_factor_var}
\title{Turns a data frame column into a factor}
\usage{
set_factor_var(df, var, reverse = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{var}{Variable name for the discrete variable to be turned into a factor}

\item{reverse}{Whether the order of the factor should be reverted, defaults
to FALSE.}
}
\value{
Data frame with factor column.
}
\description{
Turns a data frame column into a factor
}
\examples{
set_factor_var(df, var, TRUE)
}
