% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data_manipulation.R
\name{set_factors}
\alias{set_factors}
\title{Creates factors for plotting the 2018 data}
\usage{
set_factors(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with the factors x, fill, fill_label and group
}
\description{
Creates factors for plotting the 2018 data
}
\examples{
\dontrun{
set_factors(df)
}
}
