% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data_manipulation.R
\name{split_up_big_figures}
\alias{split_up_big_figures}
\title{Split up data for large figures}
\usage{
split_up_big_figures(df, max_row_limit = 30)
}
\arguments{
\item{df}{Data frame}

\item{max_row_limit}{The maximum number of rows for each figure sort key}
}
\value{
Data frame with newly assigned figure sort keys
}
\description{
This checks whether the total number of rows for a figures exceeds a limit.
If it does, the variables of the figure will be split in pieces by assigning
an additional figure_sort_key. The size of a figure is calculated by the
number of unique x-Axis + Facet variables. The function currently creates
problems by doubling headings.
}
\examples{
split_up_big_figures(df)
}
