% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_styles.R
\name{theme_rub}
\alias{theme_rub}
\title{Add RUB theme to ggplot chart}
\usage{
theme_rub(
  base_size = 11,
  base_family = "RubFlama",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  color = RUB_colors["blue"],
  facet_headings = FALSE,
  y_axis_label = FALSE
)
}
\arguments{
\item{base_size}{base font size, defaults to 11}

\item{base_family}{base font family, defaults to RubFlama}

\item{base_line_size}{base size for line elements, defaults to base_size/22}

\item{base_rect_size}{base size for rect elements, defaults to base_size/22}

\item{color}{Color for font and borders, defaults to `RUB_colors["blue"]`,
i.e. #003560.}

\item{facet_headings}{Boolean indicating whether facet headings are required,
defaults to FALSE.}

\item{y_axis_label}{Boolean indicating whether there is a label for the
y-axis, defaults to FALSE.}
}
\description{
This function allows you to add the RUB theme to your ggplot charts.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
  geom_line(colour = "#007f7f", size = 1) +
  geom_hline(yintercept = 0, size = 1, colour = "#333333") +
  theme_rub()
}
\keyword{theme_rub}
