% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_templates.R
\name{tpl_get_figure_df}
\alias{tpl_get_figure_df}
\title{Returns a character vector for the code chunk to retrieve the figure data
frame}
\usage{
tpl_get_figure_df(
  chunk_label,
  function_params,
  function_call = "dplyr::filter"
)
}
\arguments{
\item{chunk_label}{Character, chunk label}

\item{function_params}{Character}

\item{function_call}{Character, defaults to \code{dplyr::filter}}
}
\value{
List of character vectors with chunk texts
}
\description{
Returns a character vector for the code chunk to retrieve the figure data
frame
}
\examples{
# Returns code chunk as list
code_chunk_list <- tpl_get_figure_df(
  chunk_label = 1L,
  function_params = "df, figure_nr == 1L"
)

# Unlist
code_chunk_vector <- unlist(
  code_chunk_list
)

# Display code chunks as they will be written to the Rmd
writeLines(
  code_chunk_vector
)

# The function is vectorized, so you can do this:
tpl_get_figure_df(
  chunk_label = c(
    1L,
    2L
  ),
  function_params = c(
    "df, figure_nr == 1L",
    "df, figure_nr == 2L"
  )
)
}
