% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_templates.R
\name{tpl_heading}
\alias{tpl_heading}
\title{Returns a character vector for the code chunk to insert a heading}
\usage{
tpl_heading(chunk_label, heading)
}
\arguments{
\item{chunk_label}{Character, chunk label}

\item{heading}{Character, heading text}
}
\value{
List of character vectors with chunk texts
}
\description{
Returns a character vector for the code chunk to insert a heading
}
\examples{
# Returns code chunk as list
code_chunk_list <- tpl_heading(
  chunk_label = 1L,
  heading = "Section 1"
)

# Unlist
code_chunk_vector <- unlist(
  code_chunk_list
)

# Display code chunks as they will be written to the Rmd
writeLines(
  code_chunk_vector
)

# The function is vectorized, so you can do this:
tpl_heading(
  chunk_label = c(
    1L,
    2L
  ),
  heading = c(
    "Section 1",
    "Section 2"
  )
)
}
