% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_templates.R
\name{tpl_subheading}
\alias{tpl_subheading}
\title{Returns a character vector for the code chunk to insert a subheading}
\usage{
tpl_subheading(chunk_label, subheading)
}
\arguments{
\item{chunk_label}{Character, chunk label}

\item{subheading}{Character, subheading text}
}
\value{
List of character vectors with chunk texts
}
\description{
Returns a character vector for the code chunk to insert a subheading
}
\examples{
# Returns code chunk as list
code_chunk_list <- tpl_subheading(
  chunk_label = 1L,
  subheading = "Subsection 1"
)

# Unlist
code_chunk_vector <- unlist(
  code_chunk_list
)

# Display code chunks as they will be written to the Rmd
writeLines(
  code_chunk_vector
)

# The function is vectorized, so you can do this:
tpl_subheading(
  chunk_label = c(
    1L,
    2L
  ),
  subheading = c(
    "Subsection 1",
    "Subsection 2"
  )
)
}
