% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{translate_df}
\alias{translate_df}
\title{Translate data frame columns from German to English}
\usage{
translate_df(df)
}
\arguments{
\item{df}{The original data set in German}
}
\value{
Translated data set
}
\description{
Translate data frame columns from German to English
}
\examples{
\dontrun{
translate_df(df)
}
}
