% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data_manipulation.R
\name{update_cohort_axis_label}
\alias{update_cohort_axis_label}
\title{Update axis label for cohort data}
\usage{
update_cohort_axis_label(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with updated column "axis_label"
}
\description{
Update axis label for cohort data
}
\examples{
update_cohort_axis_label(df)
}
