% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_add_columns.R
\name{add_figure_height}
\alias{add_figure_height}
\title{Add column \code{figure_height} to data frame}
\usage{
add_figure_height(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with additional column \code{figure_height}
}
\description{
Add column \code{figure_height} to data frame
}
\examples{
\dontrun{
add_figure_height(df)
}
}
