% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_add_columns.R
\name{add_label_position_typ_2}
\alias{add_label_position_typ_2}
\title{Add centered y-coordinates for filtered data labels of vertical stacked bar
charts that are scaled to 100%}
\usage{
add_label_position_typ_2(df, x, y, facet)
}
\arguments{
\item{df}{Data frame}

\item{x}{x-coordinate}

\item{y}{y-coordinate}

\item{facet}{facet variable}
}
\value{
Data frame with additional column "label_" + the name of the
   y-coordinate variable
}
\description{
Calculates the y-coordinates for vertical stacked bar charts, centered for
   each group. The default function does not work if some labels are
   filtered.
}
\examples{
x <- c("WiSe 13/14", "WiSe 13/14", "WiSe 13/14", "WiSe 13/14")
y <- c(1989, 58, 163, 470)
fill <- c("Bachelor 2-Fächer", "Master 1-Fach", "Master 2-Fächer", "Master of Education")
filter_cutoff <- 0.05
df <- get_label_position(df, x, y)
ggplot2::ggplot() +
  ggplot2::geom_bar(aes(x = x, y = y, fill = fill), stat = "identity") +
  ggplot2::geom_label(data = df, aes(x = x, y = y_label, group = fill, label = y))
}
