% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_bind_rows.R
\name{bind_2018_fg}
\alias{bind_2018_fg}
\title{Bind rows for \code{report_id} "2018_FG" to data frame}
\usage{
bind_2018_fg(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with additional rows for \code{report_id} "2018_FG"
}
\description{
Bind rows for \code{report_id} "2018_FG" to data frame
}
\examples{
\dontrun{
add_2018_fg(df)
}
}
