% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_example}
\alias{df_example}
\title{Example data set illustrating the use of the RUBer parameterized
reporting package, containing the data to generate all figures.}
\format{
A data frame with 21 rows and 6 variables:
\describe{
\item{report_nr}{Integer, report number used for filtering and joining}
\item{figure_nr}{Integer, unique identifier for each figure}
\item{report_type_id}{Character, one of \code{c("STG", "MED", "M_ED", "FGR", "SZMA")}.
This ID is used for filtering and conditional logic.}
\item{x}{Character, x axis values. Will only be used for ordering if
x_label is filled.}
\item{x_label}{Character, x axis values to be displayed}
\item{y}{Character, y axis value of the figure}
\item{y_axis_label}{Character, y axis label of the figure}
\item{fill}{Integer, values used for ordering of the fill}
\item{fill_label}{Character, values used for the fill labels}
\item{facet}{Character, values to use for facetting the figure}
\item{group}{Integer, values to use for the grouping}
\item{group_label}{Character, values to use as group labels}
\item{source_caption}{Character, the source of the figure's data, displayed
at the bottom-right of each figure. Used as caption argument in the
call to \code{ggplot2::labs}.}
\item{question_txt}{Character, the question posed to the survey's
respondents and displayed as figure title}
\item{figure_type_id}{Integer, ID determining the type of figure to plot.
Currently one of four: 1 - stacked bar chart; 2 - vertical stacked
bar chart scaled to 100\%; 3 - horizontal stacked bar chart scaled
to 100\%, 4 - line chart.}
\item{figure_caption}{Character, the figure caption displayed above each
figure.}
\item{heading}{Character, the level 1 heading appearing above this figure,
only displayed when the corresponding boolean is true.}
\item{subheading}{Character, the level 2 heading appearing above this
figure, only displayed when the corresponding boolean is true.}
\item{is_heading}{Boolean, determines whether the level 1 heading is
printed above the figure.}
\item{is_subheading}{Boolean, determines whether the level 2 heading is
printed above the figure.}
\item{report_author}{Character, the report author, appearing on the title
page, on  the header of each page.}
\item{report_title}{Character, the report title, appearing on the title
page and in the header of each page.}
\item{file_name}{Character, the file name for this report}
\item{figure_height}{Numeric, height of the figure in inches}
}
}
\usage{
df_example
}
\description{
All data values in this data set are algorithmically generated. With that
exception, though, the data set closely resembles the confidential data the
package was developed for. It contains the data to generate
0 reports with varying numbers
of figures each.
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab df_example \cr
   Number of rows \tab 21 \cr
   Number of columns \tab 6 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 5 \cr
   numeric \tab 1 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   studieneingang \tab 0 \tab 1.00 \tab 5 \tab 58 \tab 0 \tab 21 \tab 0 \cr
   koepfe_rub \tab 4 \tab 0.81 \tab 1 \tab 6 \tab 0 \tab 17 \tab 0 \cr
   koepfe_rub_perc \tab 4 \tab 0.81 \tab 3 \tab 5 \tab 0 \tab 16 \tab 0 \cr
   koepfe_bef \tab 4 \tab 0.81 \tab 1 \tab 5 \tab 0 \tab 17 \tab 0 \cr
   koepfe_bef_perc \tab 4 \tab 0.81 \tab 3 \tab 5 \tab 0 \tab 16 \tab 0 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   row_id \tab 0 \tab 1 \tab 11 \tab 6.2 \tab 1 \tab 6 \tab 11 \tab 16 \tab 21 \tab ▇▆▆▆▆ \cr
}
}
\keyword{datasets}
