% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{filter_report}
\alias{filter_report}
\title{Filter data frame based on \code{report_nr} column}
\usage{
filter_report(df, report_nr)
}
\arguments{
\item{df}{Data frame}

\item{report_nr}{Required integer indicating the report_nr}
}
\value{
Filtered data frame with data for the report nr
}
\description{
Filter data frame based on \code{report_nr} column
}
\examples{
filter_report(df_fake, report_nr = 12)
}
