% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_author}
\alias{get_author}
\title{Get author for report title page}
\usage{
get_author(param_list)
}
\arguments{
\item{param_list}{List of parameters containing report_id, faculty_txt,
subject_txt, subject_area_txt}
}
\value{
String
}
\description{
Get author for report title page
}
\examples{
\dontrun{
get_author(df)
}
}
