% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data_manipulation.R,
%   R/wrangle_get_functions.R
\name{get_figure_caption}
\alias{get_figure_caption}
\title{Get figure caption}
\usage{
get_figure_caption(param_list)

get_figure_caption(param_list)
}
\arguments{
\item{df}{Dataframe bzw. Tibble}

\item{df}{Dataframe bzw. Tibble}
}
\value{
Caption as character

Caption as character
}
\description{
Get figure caption

Get figure caption
}
\examples{
get_figure_caption(df)
get_figure_caption(df)
}
