% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_file_name}
\alias{get_file_name}
\title{Get file name for automatic report generation}
\usage{
get_file_name(param_list)
}
\arguments{
\item{param_list}{}
}
\value{
String
}
\description{
Get file name for automatic report generation
}
\examples{
get_file_name(param_list)
}
