% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot.R
\name{get_label_formula}
\alias{get_label_formula}
\title{Get formula for calculating position of value labels}
\usage{
get_label_formula(label_var, is_percentage = FALSE)
}
\arguments{
\item{label_var}{The name of the variable requiring value labels}

\item{is_percentage}{Optional boolean indicating whether the value label is
expressed in absolute numbers or as a percentage, defaults to false}
}
\value{
A defused expression for calculating the position of the y-label
}
\description{
Get formula for calculating position of value labels
}
\examples{
\dontrun{
get_label_formula(y_var = cyl, label_reverse = TRUE, is_percentage = FALSE)
}
}
