% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot.R
\name{get_legend_columns}
\alias{get_legend_columns}
\title{Gets number of legend columns based on the plot width and the text base size}
\usage{
get_legend_columns(
  legend_text,
  y_axis_text,
  legend_key_width = plot_width/100,
  legend_key_spacing = plot_width/100,
  plot_width = 6.8,
  base_size = 11,
  base_family = get_font_df()[["family"]]
)
}
\arguments{
\item{legend_text}{Vector with the legend text}

\item{y_axis_text}{Vector with the text labels of the y axis}

\item{legend_key_width}{Legend key width}

\item{legend_key_spacing}{Legend key spacing}

\item{plot_width}{Width of the plot in inches, defaults to 6.8}

\item{base_size}{base font size, defaults to 11}

\item{base_family}{base font family, defaults to RubFlama}
}
\value{
Numeric with the number of columns for the legend
}
\description{
Gets number of legend columns based on the plot width and the text base size
}
\examples{
get_legend_columns(
  legend_text = c(
    "1 - eigener Verdienst",
    "2 - Mittel der Eltern/Verwandten",
    "3 - Förderung nach BAföG",
    "4 - Stipendium",
    "5 - Sonstiges"
  ),
  y_axis_text = c(
    "Bachelor 2-Fächer (n=251)",
    "FG Bachelor 2-Fächer (n=1.310)"
  )
)
}
