% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_colors.R
\name{get_RUB_palettes}
\alias{get_RUB_palettes}
\title{Return function to interpolate a RUB color palette}
\usage{
get_RUB_palettes(palette = "discrete", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to
\code{\link[grDevices]{colorRamp}}}
}
\value{
Color palette
}
\description{
Return function to interpolate a RUB color palette
}
\examples{
get_RUB_palettes(palette = "continuous")
}
