% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_bind_rows.R
\name{get_subject_group_lhs}
\alias{get_subject_group_lhs}
\title{Get left hand side of generated subject group data}
\usage{
get_subject_group_lhs(df)
}
\arguments{
\item{df}{}
}
\value{
Data frame with left hand side of generated subject group data
}
\description{
Get left hand side of generated subject group data
}
\examples{
get_subject_group_lhs(df)
}
