% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_title}
\alias{get_title}
\title{Get title for report title page based on the column "report_nr"}
\usage{
get_title(report_nr)
}
\arguments{
\item{report_nr}{}
}
\value{
String
}
\description{
Get title for report title page based on the column "report_nr"
}
\examples{
get_title(df)
}
