% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data_manipulation.R
\name{mutate_plot_variables}
\alias{mutate_plot_variables}
\title{Generiert für die Abbildungstypen 1-4 Variablen für die Ploterstellung}
\usage{
mutate_plot_variables(df)
}
\arguments{
\item{df}{Dataframe bzw. Tibble}
}
\value{
Dataframe bzw. Tibble mit den zusätzlichen Spalten x, y, fill, fill_label, fill_reverse, facet und group.
}
\description{
Generiert für die Abbildungstypen 1-4 Variablen für die Ploterstellung
}
\examples{
mutate_plot_variables(df_filtered)
}
