% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot.R
\name{plot_figure}
\alias{plot_figure}
\title{Plot RUB figure}
\usage{
plot_figure(df, font_family = get_font_df()[["family"]])
}
\arguments{
\item{df}{Data Frame}

\item{font_family}{Character, the font family to use for all plots, defaults
to \code{get_font_df()[["family"]]}}
}
\value{
ggplot object
}
\description{
Plot RUB figure
}
\examples{
df_example \%>\%
  dplyr::filter(
    .data[["report_nr"]] == 6L,
    .data[["figure_count"]] == 1L
  ) \%>\%
  RUBer::plot_figure()
}
