% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{register_font_df}
\alias{register_font_df}
\title{Register font contained in font data frame}
\usage{
register_font_df(font_df = RUBer::get_font_df())
}
\arguments{
\item{font_df}{Data frame with one row obtained by \code{RUBer::get_font_df()}}
}
\value{
Invisibly returns font family
}
\description{
Register font contained in font data frame
}
\examples{
register_font_df()
}
