% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{render_report}
\alias{render_report}
\title{Render report}
\usage{
render_report(
  p_df,
  p_df_stg = NULL,
  report_nr,
  rmd_template = here::here("datenreport_new.Rmd"),
  output_directory = NULL,
  output_filename = NULL,
  date = format(Sys.Date(), format = "\%B \%Y"),
  post_process = TRUE
)
}
\arguments{
\item{p_df}{Data frame containing the data for all reports}

\item{p_df_stg}{Optional data frame with information on cases}

\item{report_nr}{Report number of the report}

\item{rmd_template}{Path to the R Markdown File for that report, defaults to
\code{here::here("datenreport_new.Rmd")}}

\item{output_directory}{Output directory for the rendered report, defaults to
\code{here::here("output")}}

\item{output_filename}{Output filename for the rendered report, defaults to
\code{p_df[[1, "file_name"]]}}

\item{date}{Date of the report displayed on the title page, defaults to
\code{format(Sys.Date(), format= "\%B \%Y")}.}

\item{post_process}{Boolean, whether \code{post_process} gets called on the
output file}
}
\description{
Render report
}
\examples{
\dontrun{
render_report(df, 12)
}
}
