% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{render_report}
\alias{render_report}
\title{Render report}
\usage{
render_report(p_df, p_report_nr)
}
\arguments{
\item{p_df}{Data frame containing the data for all reports}

\item{report_nr}{Report number for the report to be rendered}
}
\description{
Render report
}
\examples{
render_report(df, 12)
}
