% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data_manipulation.R
\name{replace_cohort_data}
\alias{replace_cohort_data}
\title{Replaces original cohort data with corrected cohort data.}
\usage{
replace_cohort_data(df, df_orga)
}
\arguments{
\item{df}{Data frame}

\item{df_orga}{Data frame with organization hierarchies}
}
\value{
Original data frame with new data for figure_id 4
}
\description{
Replaces original cohort data with corrected cohort data.
}
\examples{
\dontrun{
replace_cohort_data(df, df_orga)
}
}
