% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{replace_placeholder}
\alias{replace_placeholder}
\title{Replace placeholder strings in docx document}
\usage{
replace_placeholder(
  doc,
  placeholder_text = "PLACEHOLDER_TAB",
  replacement_text = "\\t"
)
}
\arguments{
\item{doc}{an rdocx object}

\item{placeholder_text}{Character with the placeholder to be replaced, defaults to "PLACEHOLDER_TAB"}

\item{replacement_text}{Character with replacement text, defaults to "\\t"}
}
\value{
Side effects
}
\description{
Replace placeholder strings in docx document
}
\examples{
\dontrun{
replace_placeholder(
   path = "test.docx",
   new_path = "test_replaced.docx"
)
}
}
\keyword{internal}
