% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot.R
\name{rub_plot_typ_1}
\alias{rub_plot_typ_1}
\title{Zeichne gestapeltes Balkendiagramm (Abbildungstyp 1)}
\usage{
rub_plot_typ_1(df, x, y, y_label = "", fill, fill_label, caption = "",
  filter_cutoff)
}
\arguments{
\item{df}{Data Frame bzw. Tibble}

\item{x}{x-Position des Balkens}

\item{y}{y-Position des Balkens}

\item{y_label}{Beschriftung der y-Achse}

\item{fill}{Gruppierungsvariable für das Stapeln der Säulen (z.B. Abschluss ID)}

\item{fill_label}{Beschriftung für die einzelnen Werte der Gruppierungsvariablen (z.B. Abschluss DTXT)}

\item{caption}{Quellenangabe für die Abbildung (Präfix 'Quelle:' wird automatisch hinzugefügt)}

\item{filter_cutoff}{Schwellwert, ab dem Datenlabels unterdrückt werden (als Bruchteil von 100)}
}
\value{
Ein ggplot Objekt
}
\description{
Gestapeltes Balkendiagramm im RUB-Corporate Design mit optionalem Label für die y-Achse, optionaler Quellenangabe und der Möglichkeit, Datenlabels zu filtern.
}
\examples{
rub_plot_typ_1(df = df_fig_t1, x = time, y = value_n_total, fill = degree_sort, fill_label = degree_txt )
}
