% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot.R
\name{rub_plot_typ_2}
\alias{rub_plot_typ_2}
\title{Zeichne gestapeltes Balkendiagramm als Anteile von 100 mit Facets (Abbildungstyp 2)}
\usage{
rub_plot_typ_2(df, x, y, fill, fill_label, fill_reverse = FALSE, facet,
  caption = "", filter_cutoff = 0.04)
}
\arguments{
\item{df}{Data Frame bzw. Tibble}

\item{x}{x-Position des Balkens}

\item{y}{y-Position des Balkens}

\item{fill}{Gruppierungsvariable für das Stapeln der Säulen (z.B. Abschluss ID)}

\item{fill_label}{Beschriftung für die einzelnen Werte der Gruppierungsvariablen (z.B. Abschluss DTXT)}

\item{fill_reverse}{Boolean, ob die Reihenfolge der Gruppierungsvariable invertiert werden soll}

\item{facet}{Name der Variable, die die Abbildung in Facets aufteilt}

\item{caption}{Quellenangabe für die Abbildung (Präfix 'Quelle:' wird automatisch hinzugefügt)}

\item{filter_cutoff}{Schwellwert, ab dem Datenlabels unterdrückt werden (als Bruchteil von 100)}
}
\value{
Ein ggplot Objekt
}
\description{
Zeichne gestapeltes Balkendiagramm als Anteile von 100 mit Facets (Abbildungstyp 2)
}
\examples{
rub_plot_typ_2(df = df_fig, x = variable_txt, y = value_percentage, fill = value_id, fill_label = value_txt, fill_reverse = df_fig[[1, "scale_invert_flag"]], facet = degree_id, caption = "Fuck You, Bitch!")
}
