% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot.R
\name{rub_plot_type_4}
\alias{rub_plot_type_4}
\title{Plot grouped line chart (figure type 4)}
\usage{
rub_plot_type_4(
  df,
  x_var,
  x_axis_label = "",
  y_var,
  y_axis_label = "",
  group_var,
  group_label = NULL,
  caption = "",
  caption_prefix = "Quelle:",
  filter_cutoff = 5,
  facet_var = NULL,
  color = RUB_colors["blue"],
  palette_reverse = FALSE,
  base_family = "RubFlama",
  base_size = 11
)
}
\arguments{
\item{df}{Data frame}

\item{x_var}{Required variable name for the variable containing the discrete
x-coordinates.}

\item{x_axis_label}{Optional label for the x-axis, defaults to an empty
string.}

\item{y_var}{Required variable name for the variable containing the
continuous y-coordinates.}

\item{y_axis_label}{Optional label for the y-axis, defaults to an empty
string.}

\item{group_var}{Variable name for the discrete variable which determines the
groups forming one line, e.g. degree_id.}

\item{group_label}{Optional variable name for the character variable
containing the names of the group variable (e.g. degree_txt), defaults to #'     NULL.}

\item{caption}{Optional character containing the data source for the figure
(prefix 'Quelle:' is automatically added).}

\item{caption_prefix}{Optional character containing the prefix for the
caption, defaults to 'Quelle:'.}

\item{filter_cutoff}{Optional integer marking the cutoff below which all
value labels are suppressed, defaults to 5.}

\item{facet_var}{Optional variable name for the discrete variable to facet
by, defaults to NULL.}

\item{color}{Color for font and borders, defaults to \code{RUB_colors["blue"]},
i.e. #003560.}

\item{palette_reverse}{Optional boolean indicating whether the colors in the
palette should be reversed, defaults to FALSE.}

\item{base_family}{base font family, defaults to RubFlama}

\item{base_size}{base font size, defaults to 11}
}
\value{
A ggplot2 object
}
\description{
Plot grouped line chart (figure type 4)
}
\examples{
\dontrun{
rub_plot_type_4(
   df = df_fig_t4,
   x = time,
   y = value_n_total,
   group = degree_sort,
   group_label = degree_txt
)
}
}
