% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_styles.R
\name{rub_style}
\alias{rub_style}
\title{Add RUB theme to ggplot chart}
\usage{
rub_style(font = "RubFlama")
}
\arguments{
\item{font}{Font for the ggplot theme, defaults to RubFlama}
}
\description{
This function allows you to add the RUB theme to your ggplot charts.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
  geom_line(colour = "#007f7f", size = 1) +
  geom_hline(yintercept = 0, size = 1, colour = "#333333") +
  rub_style()
}
\keyword{rub_style}
