% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_styles.R
\name{rub_style_flextable}
\alias{rub_style_flextable}
\title{Applies RUB-Style to flextable object}
\usage{
rub_style_flextable(tbl, font = "RUB Scala TZ")
}
\arguments{
\item{tbl}{Flextable object}

\item{font}{Font for the ggplot theme, defaults to RUB Scala TZ}
}
\value{
Styled flextable object
}
\description{
Applies RUB-Style to flextable object
}
\examples{
rub_style_flextable(tbl)
}
