% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rub_plot.R
\name{set_factor_var}
\alias{set_factor_var}
\title{Prepares var and var_label columns for plotting by ordering them and turning
them into factors}
\usage{
set_factor_var(df, var, var_label = NULL, reverse = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{var}{Required variable name for the discrete variable to be turned into
a factor. This variable is sorted alphabetically to determine the order.}

\item{var_label}{Optional variable name for the discrete variable labels
to be used instead of var. If var is a sort key, for instance, var_label
can be used for the actual labels.}

\item{reverse}{Whether the order of the factor should be reverted, defaults
to FALSE.}
}
\value{
Data frame with factor column.
}
\description{
Prepares var and var_label columns for plotting by ordering them and turning
them into factors
}
\examples{
\dontrun{
set_factor_var(df, var, TRUE)
}
}
