% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data_manipulation.R
\name{set_factors}
\alias{set_factors}
\title{Erzeuge Faktoren für die Abbildungserstellung}
\usage{
set_factors(df)
}
\arguments{
\item{df}{Dataframe bzw. Tibble}
}
\value{
Dataframe bzw. Tibble mit den Faktoren x, fill, fill_label und group
}
\description{
Erzeuge Faktoren für die Abbildungserstellung
}
\examples{
set_factors(df)
}
