% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sort_report}
\alias{sort_report}
\title{Sorts the report data frame}
\usage{
sort_report(df)
}
\arguments{
\item{df}{Data frame with report data}
}
\value{
Sorted data frame
}
\description{
Sorts report_nr, figure_sort, subject_id, subject_area_id, degree_sort,
time, variable_id, subject_group_id, value_id
}
\examples{
sort_report(df)
}
